Scripts/requirements.txt
pandas==1.2.2
Pillow==9.1.0
prefetch_generator==1.0.1
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
